Created the dev branch, and started code
The files currently contain code for datasets. At the moment we can load
a dataset, and destroy it.

The loading function needs to be expanded to do more checks as we are
are allocating a lot of memory dynamically, but no checks are being done
to see if they are actually being allocated (i.e. != NULL)