Finished initial versions of helper functions
The functions used to calculate a set of outputs from a single run
have been written, (but not yet tested). Preliminary compilation
shows no errors.

In addition, the makefile was updated to link the math library so
that exp() can be used for the sigmoid function.

The runNetwork(...) function still needs to be written.