Preliminary runNetwork function done
The runNetwork(...) function has been written such that a dataset
can be passed to the network and each member within the set will
be passed through the network.

In addition, code has been included which allows the function to
print the inputs and a combination of outputs, targets and errors
for each data member.

Again, the code compiles successfully, but has yet to be tested.

The next phase will be to implement code to build the network, followed
by the functions necessary to train the network.