Neural Network test (eventually) successful
The makefile was updated to include the debugging flag during testing.
It will remain there until a neural network built with the library
can learn successfully.

The library code has been updated to correct minor mistakes in the
code (such as missing ; and brackets), and also to include functionality
which allows the network to be run on data.

A preliminary test with the XOR problem and a set of known weights and
results show that a network will behave as expected.

NOTE: This version is still buggy and not fully complete.